Convert `with-custom-server` server file to typescript
